Add option to keep SI compliance
Converts composite units to (string) SI units when multiplying or dividing
DataArrays/Units for consistency to avoid additional scaling factors in
the "units" attribute, e.g. when converting from cm-3 to m-3.
Adds a keyword option to `init_units()` to make that behaviour optional
as it involves an additional multiplication and changes the values.