Always return `astropy.Unit` when inferring units
Relying on the `.unit` attribute to be an `astropy.Unit` fails when this
is set by some other package. It looks like at least `pandas` uses that
attribute as a string, and then using `.to()` fails.
Converts that attribute also to an `astropy.Unit` as we do in all the
other cases.