Simplify conversion for `add` and `sub`
The idea is to convert the second array to the units of the first array,
thus we only need the conversion factor. Since we made sure that
`_get_unit()` returns an `astropy.Unit`, we use `.to()` to get this
factor, or an exception when the units are not compatible.
Fixes addition and subtraction with primitive types such as floats or
ints, or `numpy.ndarray`s which typically don't have units.