Remove implicit conversion to SI units
We implicitly converted to SI units after multiplication and division.
However, this way we can end up with strange factors in the unit, and
it might not be desired at all.
Removes this implicit conversion and leaves it to the caller to convert
to practical units. Also, `.to_si_units()` can then be used after all
computations and for serializing the data to, e.g. netcdf.