Always use strings for the units attribute
Sets the "units" attribute to the string version of the units after
multiplication and division, and when explicitly converting.
Since we made sure that `_get_unit()` returns an `astropy.Unit`
instance, this was promoted to the variable's "units" attribute.
But then serializing to netcdf fails, and using the string version
for the attribute helps in those cases.